Builds and copies swiftbridge framework
Automates the process of building the swiftbridge framework
using xcodebuild and copying it to the output directory.

This eliminates the need for manual building and copying,
ensuring that the framework is always up-to-date when the
rustbridge_tests crate is built.

It also removes the previous implementation, which relied on
inferring the framework directory via xcode build settings.